BlePer/pyproject.toml
[project]
name = "bleper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"pyobjc-core>=10.1",
	"pyobjc-framework-Cocoa>=10.1",
	"pyobjc-framework-CoreBluetooth>=10.1",
	"fastapi>=0.110",
	"python-multipart>=0.0.9",
	"uvicorn[standard]>=0.27",
	"pillow>=10.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["ble_peripheral", "http_image_server", "main"]
